requirements.txt
amqp==2.2.1
asn1crypto==0.22.0
Babel==2.5.0
backports-abc==0.5
bcrypt==3.1.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
Django==1.11.4
django-celery-beat==1.0.1
enum34==1.1.6
Fabric==1.13.2
flower==0.9.2
futures==3.1.1
idna==2.6
ipaddress==1.0.18
kombu==4.1.0
lxml==3.8.0
paramiko==2.2.1
pyasn1==0.3.2
pycparser==2.18
PyJWT==1.5.2
PyNaCl==1.1.2
pytz==2017.2
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.1
urllib3==1.22
vine==1.1.4
wheel==0.24.0
